requirements.txt
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.4
Django
django-graphql-jwt==0.4.0
django-oauth-toolkit==2.3.0
django-postgresql==0.0.3
djangorestframework==3.14.0
elastic-transport==8.12.0
elasticsearch==8.12.1
flasgger==0.9.7.1
Flask==3.0.2
Flask-JWT-Extended==4.6.0
flask-restplus==0.13.0
flask-swagger-ui==4.11.1
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwcrypto==1.5.4
kombu==5.3.5
MarkupSafe==2.1.5
mistune==3.0.2
oauthlib==3.2.2
packaging==23.2
promise==2.3
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
